Chat Update and Query
