+++ 3-PB/mvp/src/serverActions/chats/getChats.ts
@@ -3,7 +3,13 @@
 import pool from "@/serverActions/utils/postgres";
 
 export async function getChats() {
-  const query = "SELECT id,title FROM chat_threads ORDER BY id DESC";
+  const query = `
+  SELECT chat_threads.id, chat_threads.title, MAX(messages.created_at) AS last_message_time
+  FROM chat_threads
+  LEFT JOIN messages ON chat_threads.id = messages.thread_id
+  GROUP BY chat_threads.id, chat_threads.title
+  ORDER BY last_message_time DESC NULLS LAST, chat_threads.id DESC;
+`;
   const { rows } = await pool.query(query);
 
   return rows.map((row: any) => ({
